pyproject.toml
[tool.poetry]
name = "crypto-trading-bot"
version = "0.1.0"
description = "This is a binance crypto-trading automatic bot."
authors = ["RomanchenkoAS <RomanchenkoAS@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
vectorbt = "^0.25.5"
jupyter = "^1.0.0"
grequests = "^0.7.0"
python-decouple = "^3.8"
python-binance = "^1.0.19"
pandas = "^2.2.0"
pandas-ta = "^0.3.14b0"
redis = "^5.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
virtualenvs.in-project = true
